refactor: narrator package
