+++ src/main/kotlin/com/jaoafa/vcspeaker/events/GoLiveStartEvent.kt
@@ -2,8 +2,8 @@ package com.jaoafa.vcspeaker.events
 
 import com.jaoafa.vcspeaker.stores.GuildStore
 import com.jaoafa.vcspeaker.tools.discord.DiscordExtensions.selfVoiceChannel
-import com.jaoafa.vcspeaker.tts.NarrationScripts
-import com.jaoafa.vcspeaker.tts.NarratorExtensions.announce
+import com.jaoafa.vcspeaker.tts.narrators.NarrationScripts
+import com.jaoafa.vcspeaker.tts.narrators.Narrator.Companion.announce
 import com.kotlindiscord.kord.extensions.extensions.Extension
 import com.kotlindiscord.kord.extensions.extensions.event
 import dev.kord.core.event.user.VoiceStateUpdateEvent
