+++ src/main/kotlin/com/jaoafa/vcspeaker/events/VoiceLeaveEvent.kt
@@ -5,8 +5,8 @@ import com.jaoafa.vcspeaker.stores.GuildStore
 import com.jaoafa.vcspeaker.tools.discord.DiscordExtensions.autoJoinEnabled
 import com.jaoafa.vcspeaker.tools.discord.DiscordExtensions.isAfk
 import com.jaoafa.vcspeaker.tools.discord.DiscordExtensions.selfVoiceChannel
-import com.jaoafa.vcspeaker.tts.NarrationScripts
-import com.jaoafa.vcspeaker.tts.NarratorExtensions.announce
+import com.jaoafa.vcspeaker.tts.narrators.NarrationScripts
+import com.jaoafa.vcspeaker.tts.narrators.Narrator.Companion.announce
 import com.kotlindiscord.kord.extensions.extensions.Extension
 import com.kotlindiscord.kord.extensions.extensions.event
 import dev.kord.core.event.user.VoiceStateUpdateEvent
